requirements.txt
Flask
opencv-python
numpy
tflite-runtime
Pillow
waitress
RPi.GPIO